requirements.txt
numpy==1.26.0
pandas==2.1.0
matplotlib==3.8.0
seaborn==0.12.2
jupyter==1.0.0
notebook==7.0.4
ipywidgets==8.0.3
scipy==1.13.0
pydantic==2.6.0
pyyaml==6.0.1
tensorboard==2.16.2
tensorboardX==2.6.2.2
scikit-learn==1.3.0
plotly==5.18.0
ipython==8.15.0
